Complete exercise 5 in foundations
Used a decendant-combinator to add styles to the div container of
paragraph. Had to use a descendant combinator referencing the type p
in order to not get white space, as the container created by the html
apparently applied background color yellow to the white space. Still
a little confused as to why the .text or the p reference made it only
affect the text, but i get it a little also