Complete CSSEXERCISE Flex 01
This was relatively easy. Centered the Red Div Box within the Blue
container. I used Justify Content on the container and align self on the
div, but I also tested if it was possible to use Justify content and
align Items on the just container (yes) or Justify content and Align
Items on the item (no).