Complete CSS Basic Flex header 02
This one was a little difficult. I was able to figure out how to get
The Logo in the center and the items spaced out a little more (UL's
require their own flex:display orientation). Pushing the links was a
little difficult. At first I used center, then margins on the left link
and right link class (first with values then with auto), however those
didn't really account for the differences in font length. Then I used
the flexbox guide and realized that spacebetween would probably be a
better set up. I did this, and was able to get the images decently
spaced out. From there I was able to push them out to the edges.
I then looked at the solution and noted that theirs was a little simpler
than mine in that they used a margin of 0 and a padding of 0 to push it
out to the edges. At first I didn't get it, but after doing some
googling I found that UL's and OL's have default padding and margin that
need to be accounted for.