Solve Header 02
Another somewhat hard one. Things I learned from this and reinforced
Set the margin and padding on your links to 0 default. Nest containers
(divs). If you have a link container, it most likely should be within
a div container. You have to add a gap to both the individual links
and also one to the container (depending) to seprate the different
items in the container. Space between is useful for splitting up a page
but only works properly if the header is properly nested in containers
with a limited amount of items. Put the general padding on the header.