Complete Css Exercise 04 Information
This one was diffficult. Had to start from scratch a couple of times
But running it slowly helped me understand what I was doing wrong
As for the HTML, Starting with the title, since it was in a div
container, I could either use align-text to center it, or use display:
flex + justify content. Same effect either way. From there I went back
to the html and started making div container. I needed one div container
for the Image/Text convo which I called items, and another div container
for the entire row of plants, which i called plants. Once i made the
containers. I used flex wrap- wrap to have the text go under the
images. I then ran text align center to center the text.
From there I went back to the plants div - I changed it to a flexbox
which changed it from a column to a row. From there it was just spacing
out the items properly. Some things I didn't realize I could do was
running a text align on the body which would have centered the entire
page.